requirements.txt
easydict
numpy
pandas
scikit-learn
pillow
tqdm
pytorch
torchvision
pytorch-lightning